requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.0
awscli==1.11.83
billiard==3.3.0.23
boto==2.48.0
botocore==1.5.46
celery==3.1.25
colorama==0.3.7
dj-database-url==0.4.2
Django==1.9.12
django-braces==1.11.0
django-celery==3.2.1
django-kombu==0.9.4
django-mail-templated==2.6.5
django-pyodbc==1.1.0
django-pyodbc-azure==1.9.12.0
django-storages==1.6.3
djrill==2.1.0
docutils==0.13.1
futures==3.0.5
gunicorn==19.6.0
honcho==1.0.1
jmespath==0.9.2
kombu==3.0.37
olefile==0.44
packaging==16.8
Pillow==3.0.0
psycopg2==2.7.2
pyasn1==0.2.3
pyodbc==4.0.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.13.0
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
wfastcgi==2.2.1
whitenoise==3.3.0
